Move semantic_diff.py to incremental
